+++ app/api/v1/admin.py
@@ -1,31 +1,51 @@
-from datetime import datetime, timezone, timedelta
-from typing import Annotated, Optional
+from typing import Annotated
 
-import secrets
-from fastapi import APIRouter, Depends, HTTPException, Query, status
-from pydantic import BaseModel
-from sqlalchemy import func
+from fastapi import APIRouter, Depends, Query
 from sqlalchemy.orm import Session
 
 from app.api.deps import get_current_admin
 from app.core.database import get_db
-from app.models.arcade import Arcade, ArcadeGame
-from app.models.game import Game
-from app.models.promo import PromoCode
-from app.models.ticket import TicketOffer
-from app.models.user import User
-
 from app.core.messages import (
     ARCADE_NOT_FOUND,
     GAME_NOT_FOUND,
     PROMO_CODE_NOT_FOUND,
     USER_NOT_FOUND,
 )
+from app.schemas.admin import (
+    ArcadeGameAssignmentRequest,
+    CreateArcadeRequest,
+    CreateGameRequest,
+    CreatePromoCodeRequest,
+    UpdatePromoCodeRequest,
+    UpdateUserTicketsRequest,
+)
+from app.services.admin_service import (
+    assign_game_to_arcade_service,
+    create_arcade_service,
+    create_game_service,
+    create_promo_code_service,
+    force_cancel_user_reservations_service,
+    get_admin_stats_service,
+    get_expiring_promo_codes_service,
+    get_user_deletion_impact_service,
+    list_deleted_arcades_service,
+    list_deleted_users_service,
+    list_promo_codes_service,
+    regenerate_arcade_api_key_service,
+    restore_arcade_service,
+    restore_user_service,
+    soft_delete_arcade_service,
+    soft_delete_game_service,
+    soft_delete_user_service,
+    toggle_promo_code_active_service,
+    update_arcade_service,
+    update_promo_code_service,
+    update_user_tickets_service,
+)
 
 router = APIRouter(
     dependencies=[Depends(get_current_admin)]
 )
-
 
 ARCADE_NOT_FOUND_RESPONSE = {
     404: {"description": ARCADE_NOT_FOUND}
@@ -44,92 +64,12 @@ USER_NOT_FOUND_RESPONSE = {
 }
 
 
-class CreateArcadeRequest(BaseModel):
-    nom: str
-    description: Optional[str] = ""
-    arcade_image: Optional[str] = None
-    localisation: str
-    latitude: float = 0.0
-    longitude: float = 0.0
-
-
-class CreateGameRequest(BaseModel):
-    nom: str
-    description: str
-    game_image: Optional[str] = None
-    min_players: int = 1
-    max_players: int = 2
-    ticket_cost: int = 1
-
-
-class CreatePromoCodeRequest(BaseModel):
-    code: str
-    tickets_reward: int
-    is_single_use_global: bool = False
-    is_single_use_per_user: bool = True
-    usage_limit: Optional[int] = None
-    valid_from: Optional[datetime] = None
-    valid_until: Optional[datetime] = None
-    is_active: bool = True
-
-
-class UpdatePromoCodeRequest(BaseModel):
-    tickets_reward: Optional[int] = None
-    is_single_use_global: Optional[bool] = None
-    is_single_use_per_user: Optional[bool] = None
-    usage_limit: Optional[int] = None
-    valid_from: Optional[datetime] = None
-    valid_until: Optional[datetime] = None
-    is_active: Optional[bool] = None
-
-
-class UpdateUserTicketsRequest(BaseModel):
-    user_id: int
-    tickets_to_add: int
-
-
-class ArcadeGameAssignmentRequest(BaseModel):
-    arcade_id: int
-    game_id: int
-    slot_number: int
-
-
-def generate_unique_arcade_api_key(db: Session) -> str:
-    """Génère une clé API unique pour une borne."""
-    new_api_key = f"arcade_key_{secrets.token_urlsafe(16)}"
-    while db.query(Arcade).filter(Arcade.api_key == new_api_key).first():
-        new_api_key = f"arcade_key_{secrets.token_urlsafe(16)}"
-    return new_api_key
-
-
-# === GESTION DES BORNES ===
 @router.post("/arcades/")
 async def create_arcade(
     arcade_data: CreateArcadeRequest,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Crée une nouvelle borne d'arcade."""
-    api_key = generate_unique_arcade_api_key(db)
-
-    arcade = Arcade(
-        nom=arcade_data.nom,
-        description=arcade_data.description,
-        arcade_image=arcade_data.arcade_image,
-        api_key=api_key,
-        localisation=arcade_data.localisation,
-        latitude=arcade_data.latitude,
-        longitude=arcade_data.longitude,
-    )
-
-    db.add(arcade)
-    db.commit()
-    db.refresh(arcade)
-
-    return {
-        "message": "Borne créée",
-        "arcade_id": arcade.id,
-        "api_key": api_key,
-    }
+    return create_arcade_service(db, arcade_data)
 
 
 @router.put("/arcades/{arcade_id}", responses=ARCADE_NOT_FOUND_RESPONSE)
@@ -138,32 +78,7 @@ async def update_arcade(
     arcade_data: CreateArcadeRequest,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Met à jour les informations d'une borne d'arcade."""
-    arcade = db.query(Arcade).filter(
-        Arcade.id == arcade_id,
-        Arcade.is_deleted.is_(False),
-    ).first()
-
-    if not arcade:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Borne non trouvée",
-        )
-
-    arcade.nom = arcade_data.nom
-    arcade.description = arcade_data.description
-    arcade.arcade_image = arcade_data.arcade_image
-    arcade.localisation = arcade_data.localisation
-    arcade.latitude = arcade_data.latitude
-    arcade.longitude = arcade_data.longitude
-
-    db.commit()
-    db.refresh(arcade)
-
-    return {
-        "message": "Borne mise à jour",
-        "arcade_id": arcade.id,
-    }
+    return update_arcade_service(db, arcade_id, arcade_data)
 
 
 @router.put(
@@ -178,188 +93,31 @@ async def assign_game_to_arcade(
     assignment: ArcadeGameAssignmentRequest,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Assigne un jeu à une borne sur un slot spécifique."""
-    if assignment.arcade_id != arcade_id:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="L'arcade_id du body doit correspondre à celui de l'URL",
-        )
-
-    arcade = db.query(Arcade).filter(
-        Arcade.id == assignment.arcade_id,
-        Arcade.is_deleted.is_(False),
-    ).first()
-    if not arcade:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Borne non trouvée",
-        )
-
-    game = db.query(Game).filter(
-        Game.id == assignment.game_id,
-        Game.is_deleted.is_(False),
-    ).first()
-    if not game:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Jeu non trouvé",
-        )
-
-    if assignment.slot_number not in (1, 2):
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Le slot doit être 1 ou 2",
-        )
-
-    existing = db.query(ArcadeGame).filter(
-        ArcadeGame.arcade_id == assignment.arcade_id,
-        ArcadeGame.slot_number == assignment.slot_number,
-    ).first()
-
-    if existing:
-        db.delete(existing)
-
-    arcade_game = ArcadeGame(
-        arcade_id=assignment.arcade_id,
-        game_id=assignment.game_id,
-        slot_number=assignment.slot_number,
-    )
-
-    db.add(arcade_game)
-    db.commit()
-
-    return {
-        "message": f"Jeu {game.nom} assigné au slot {assignment.slot_number} de la borne {arcade.nom}"
-    }
+    return assign_game_to_arcade_service(db, arcade_id, assignment)
 
 
-# === GESTION DES JEUX ===
 @router.post("/games/")
 async def create_game(
     game_data: CreateGameRequest,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Crée un nouveau jeu."""
-    game = Game(
-        nom=game_data.nom,
-        description=game_data.description,
-        game_image=game_data.game_image,
-        min_players=game_data.min_players,
-        max_players=game_data.max_players,
-        ticket_cost=game_data.ticket_cost,
-    )
-
-    db.add(game)
-    db.commit()
-    db.refresh(game)
-
-    return {
-        "message": "Jeu créé",
-        "game_id": game.id,
-    }
+    return create_game_service(db, game_data)
 
 
 @router.delete("/games/{game_id}")
 async def soft_delete_game(
-        game_id: int,
-        db: Session = Depends(get_db),
-        _: dict = Depends(get_current_admin)
+    game_id: int,
+    db: Annotated[Session, Depends(get_db)],
 ):
-    """Supprime un jeu (soft delete)."""
-
-    game = db.query(Game).filter(
-        Game.id == game_id,
-        Game.is_deleted == False
-    ).first()
-
-    if not game:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Jeu non trouvé"
-        )
-
-    # Bloquer si des réservations actives utilisent ce jeu
-    from app.models.reservation import Reservation, ReservationStatus
-    active_reservations = db.query(Reservation).filter(
-        Reservation.game_id == game_id,
-        Reservation.status.in_([ReservationStatus.WAITING, ReservationStatus.PLAYING]),
-        Reservation.is_deleted == False
-    ).count()
-
-    if active_reservations > 0:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Impossible de supprimer le jeu : {active_reservations} réservation(s) active(s)"
-        )
-
-    # Soft delete du jeu
-    game.is_deleted = True
-    game.deleted_at = datetime.now(timezone.utc)
-
-    # Soft delete des associations arcade-jeu
-    arcade_games = db.query(ArcadeGame).filter(
-        ArcadeGame.game_id == game_id,
-        ArcadeGame.is_deleted == False
-    ).all()
-
-    for ag in arcade_games:
-        ag.is_deleted = True
-        ag.deleted_at = datetime.now(timezone.utc)
-
-    db.commit()
-
-    return {
-        "message": f"Jeu '{game.nom}' supprimé avec succès",
-        "game_id": game.id,
-        "deleted_associations": len(arcade_games)
-    }
+    return soft_delete_game_service(db, game_id)
 
 
-# === GESTION DES CODES PROMO ===
 @router.post("/promo-codes/")
 async def create_promo_code(
     promo_data: CreatePromoCodeRequest,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Crée un nouveau code promo avec gestion des dates."""
-    if promo_data.valid_from and promo_data.valid_until:
-        if promo_data.valid_until <= promo_data.valid_from:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="La date d'expiration doit être après la date de début",
-            )
-
-    existing = db.query(PromoCode).filter(
-        PromoCode.code == promo_data.code.upper().strip()
-    ).first()
-
-    if existing:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Ce code promo existe déjà",
-        )
-
-    promo_code = PromoCode(
-        code=promo_data.code.upper().strip(),
-        tickets_reward=promo_data.tickets_reward,
-        is_single_use_global=promo_data.is_single_use_global,
-        is_single_use_per_user=promo_data.is_single_use_per_user,
-        usage_limit=promo_data.usage_limit,
-        valid_from=promo_data.valid_from,
-        valid_until=promo_data.valid_until,
-        is_active=promo_data.is_active,
-    )
-
-    db.add(promo_code)
-    db.commit()
-    db.refresh(promo_code)
-
-    return {
-        "message": "Code promo créé",
-        "promo_code_id": promo_code.id,
-        "is_valid_now": promo_code.is_valid_now(),
-        "days_until_expiry": promo_code.days_until_expiry(),
-    }
+    return create_promo_code_service(db, promo_data)
 
 
 @router.put("/promo-codes/{promo_code_id}", responses=PROMO_CODE_NOT_FOUND_RESPONSE)
@@ -368,48 +126,7 @@ async def update_promo_code(
     update_data: UpdatePromoCodeRequest,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Met à jour un code promo existant."""
-    promo_code = db.query(PromoCode).filter(
-        PromoCode.id == promo_code_id,
-        PromoCode.is_deleted.is_(False),
-    ).first()
-
-    if not promo_code:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Code promo non trouvé",
-        )
-
-    valid_from = (
-        update_data.valid_from
-        if update_data.valid_from is not None
-        else promo_code.valid_from
-    )
-    valid_until = (
-        update_data.valid_until
-        if update_data.valid_until is not None
-        else promo_code.valid_until
-    )
-
-    if valid_from and valid_until and valid_until <= valid_from:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="La date d'expiration doit être après la date de début",
-        )
-
-    update_dict = update_data.model_dump(exclude_unset=True)
-    for field, value in update_dict.items():
-        setattr(promo_code, field, value)
-
-    db.commit()
-    db.refresh(promo_code)
-
-    return {
-        "message": "Code promo mis à jour",
-        "promo_code_id": promo_code.id,
-        "is_valid_now": promo_code.is_valid_now(),
-        "days_until_expiry": promo_code.days_until_expiry(),
-    }
+    return update_promo_code_service(db, promo_code_id, update_data)
 
 
 @router.get("/promo-codes/")
@@ -417,39 +134,7 @@ async def list_promo_codes(
     db: Annotated[Session, Depends(get_db)],
     include_expired: bool = False,
 ):
-    """Liste tous les codes promo avec filtrage optionnel."""
-    query = db.query(PromoCode).filter(PromoCode.is_deleted.is_(False))
-
-    if not include_expired:
-        now = datetime.now(timezone.utc)
-        query = query.filter(
-            PromoCode.valid_until.is_(None) | (PromoCode.valid_until > now)
-        )
-
-    promo_codes = query.order_by(PromoCode.created_at.desc()).all()
-
-    result = []
-    for promo in promo_codes:
-        result.append(
-            {
-                "id": promo.id,
-                "code": promo.code,
-                "tickets_reward": promo.tickets_reward,
-                "usage_limit": promo.usage_limit,
-                "current_uses": promo.current_uses,
-                "is_single_use_global": promo.is_single_use_global,
-                "is_single_use_per_user": promo.is_single_use_per_user,
-                "valid_from": promo.valid_from.isoformat() if promo.valid_from else None,
-                "valid_until": promo.valid_until.isoformat() if promo.valid_until else None,
-                "is_active": promo.is_active,
-                "is_valid_now": promo.is_valid_now(),
-                "is_expired": promo.is_expired(),
-                "days_until_expiry": promo.days_until_expiry(),
-                "created_at": promo.created_at.isoformat(),
-            }
-        )
-
-    return result
+    return list_promo_codes_service(db, include_expired)
 
 
 @router.post(
@@ -460,27 +145,7 @@ async def toggle_promo_code_active(
     promo_code_id: int,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Active/désactive manuellement un code promo."""
-    promo_code = db.query(PromoCode).filter(
-        PromoCode.id == promo_code_id,
-        PromoCode.is_deleted.is_(False),
-    ).first()
-
-    if not promo_code:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Code promo non trouvé",
-        )
-
-    promo_code.is_active = not promo_code.is_active
-    db.commit()
-
-    return {
-        "message": f"Code promo {'activé' if promo_code.is_active else 'désactivé'}",
-        "promo_code_id": promo_code.id,
-        "is_active": promo_code.is_active,
-        "is_valid_now": promo_code.is_valid_now(),
-    }
+    return toggle_promo_code_active_service(db, promo_code_id)
 
 
 @router.get("/promo-codes/expiring-soon")
@@ -488,83 +153,22 @@ async def get_expiring_promo_codes(
     db: Annotated[Session, Depends(get_db)],
     days_ahead: Annotated[int, Query(ge=1)] = 7,
 ):
-    """Récupère les codes promo qui expirent bientôt."""
-    now = datetime.now(timezone.utc)
-    future_date = now + timedelta(days=days_ahead)
-
-    expiring_codes = db.query(PromoCode).filter(
-        PromoCode.is_deleted.is_(False),
-        PromoCode.is_active.is_(True),
-        PromoCode.valid_until.isnot(None),
-        PromoCode.valid_until <= future_date,
-        PromoCode.valid_until > now,
-    ).order_by(PromoCode.valid_until).all()
-
-    result = []
-    for promo in expiring_codes:
-        result.append(
-            {
-                "id": promo.id,
-                "code": promo.code,
-                "tickets_reward": promo.tickets_reward,
-                "valid_until": promo.valid_until.isoformat(),
-                "days_until_expiry": promo.days_until_expiry(),
-                "current_uses": promo.current_uses,
-                "usage_limit": promo.usage_limit,
-            }
-        )
-
-    return {
-        "expiring_codes": result,
-        "total_count": len(result),
-        "days_ahead": days_ahead,
-    }
+    return get_expiring_promo_codes_service(db, days_ahead)
 
 
-# === GESTION DES UTILISATEURS ===
 @router.put("/users/tickets", responses=USER_NOT_FOUND_RESPONSE)
 async def update_user_tickets(
     update_data: UpdateUserTicketsRequest,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Ajoute ou retire des tickets à un utilisateur."""
-    user = db.query(User).filter(
-        User.id == update_data.user_id,
-        User.is_deleted.is_(False),
-    ).first()
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=USER_NOT_FOUND,
-        )
-
-    old_balance = user.tickets_balance
-    user.tickets_balance += update_data.tickets_to_add
-
-    if user.tickets_balance < 0:
-        user.tickets_balance = 0
-
-    db.commit()
-
-    return {
-        "message": f"Solde mis à jour pour {user.pseudo}",
-        "old_balance": old_balance,
-        "new_balance": user.tickets_balance,
-        "tickets_added": update_data.tickets_to_add,
-    }
+    return update_user_tickets_service(db, update_data)
 
 
 @router.get("/users/deleted")
 async def list_deleted_users(
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Liste les utilisateurs supprimés (soft delete)."""
-    deleted_users = db.query(User).filter(
-        User.is_deleted.is_(True)
-    ).all()
-
-    return deleted_users
+    return list_deleted_users_service(db)
 
 
 @router.put("/users/{user_id}/restore", responses=USER_NOT_FOUND_RESPONSE)
@@ -572,26 +176,7 @@ async def restore_user(
     user_id: int,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Restaure un utilisateur supprimé."""
-    user = db.query(User).filter(User.id == user_id).first()
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=USER_NOT_FOUND,
-        )
-
-    if not user.is_deleted:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Cet utilisateur n'est pas supprimé",
-        )
-
-    user.is_deleted = False
-    user.deleted_at = None
-    db.commit()
-
-    return {"message": f"Utilisateur {user.pseudo} restauré"}
+    return restore_user_service(db, user_id)
 
 
 @router.delete("/users/{user_id}", responses=USER_NOT_FOUND_RESPONSE)
@@ -599,86 +184,7 @@ async def soft_delete_user(
     user_id: int,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Supprime un utilisateur (soft delete) - Accès admin uniquement."""
-    user = db.query(User).filter(
-        User.id == user_id,
-        User.is_deleted.is_(False),
-    ).first()
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=USER_NOT_FOUND,
-        )
-
-    from app.models.friend import Friendship
-    from app.models.promo import PromoUse
-    from app.models.reservation import Reservation, ReservationStatus
-    from app.models.ticket import TicketPurchase
-
-    active_reservations = db.query(Reservation).filter(
-        (Reservation.player_id == user_id) | (Reservation.player2_id == user_id),
-        Reservation.status.in_([ReservationStatus.WAITING, ReservationStatus.PLAYING]),
-        Reservation.is_deleted.is_(False),
-    ).count()
-
-    if active_reservations > 0:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=(
-                f"Impossible de supprimer l'utilisateur : {active_reservations} réservation(s) active(s). "
-                "Veuillez d'abord gérer les réservations en cours."
-            ),
-        )
-
-    now = datetime.now(timezone.utc)
-
-    user.is_deleted = True
-    user.deleted_at = now
-
-    friendships = db.query(Friendship).filter(
-        (Friendship.requester_id == user_id) | (Friendship.requested_id == user_id),
-        Friendship.is_deleted.is_(False),
-    ).all()
-
-    deleted_friendships = 0
-    for friendship in friendships:
-        friendship.is_deleted = True
-        friendship.deleted_at = now
-        deleted_friendships += 1
-
-    promo_uses = db.query(PromoUse).filter(
-        PromoUse.user_id == user_id,
-        PromoUse.is_deleted.is_(False),
-    ).all()
-
-    deleted_promo_uses = 0
-    for promo_use in promo_uses:
-        promo_use.is_deleted = True
-        promo_use.deleted_at = now
-        deleted_promo_uses += 1
-
-    ticket_purchases = db.query(TicketPurchase).filter(
-        TicketPurchase.user_id == user_id,
-        TicketPurchase.is_deleted.is_(False),
-    ).all()
-
-    deleted_purchases = 0
-    for purchase in ticket_purchases:
-        purchase.is_deleted = True
-        purchase.deleted_at = now
-        deleted_purchases += 1
-
-    db.commit()
-
-    return {
-        "message": f"Utilisateur '{user.pseudo}' supprimé avec succès",
-        "user_id": user.id,
-        "deleted_friendships": deleted_friendships,
-        "deleted_promo_uses": deleted_promo_uses,
-        "deleted_purchases": deleted_purchases,
-        "note": "Les scores sont conservés de manière anonymisée pour l'intégrité des données de jeu",
-    }
+    return soft_delete_user_service(db, user_id)
 
 
 @router.get("/users/{user_id}/deletion-impact", responses=USER_NOT_FOUND_RESPONSE)
@@ -686,93 +192,7 @@ async def get_user_deletion_impact(
     user_id: int,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Analyse l'impact de la suppression d'un utilisateur avant de la confirmer."""
-    from app.models.friend import Friendship
-    from app.models.promo import PromoUse
-    from app.models.reservation import Reservation, ReservationStatus
-    from app.models.score import Score
-    from app.models.ticket import TicketPurchase
-
-    user = db.query(User).filter(
-        User.id == user_id,
-        User.is_deleted.is_(False),
-    ).first()
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Utilisateur non trouvé",
-        )
-
-    active_reservations = db.query(Reservation).filter(
-        (Reservation.player_id == user_id) | (Reservation.player2_id == user_id),
-        Reservation.status.in_([ReservationStatus.WAITING, ReservationStatus.PLAYING]),
-        Reservation.is_deleted.is_(False),
-    ).count()
-
-    completed_reservations = db.query(Reservation).filter(
-        (Reservation.player_id == user_id) | (Reservation.player2_id == user_id),
-        Reservation.status.in_([ReservationStatus.COMPLETED, ReservationStatus.CANCELLED]),
-        Reservation.is_deleted.is_(False),
-    ).count()
-
-    friendships_count = db.query(Friendship).filter(
-        (Friendship.requester_id == user_id) | (Friendship.requested_id == user_id),
-        Friendship.is_deleted.is_(False),
-    ).count()
-
-    promo_uses_count = db.query(PromoUse).filter(
-        PromoUse.user_id == user_id,
-        PromoUse.is_deleted.is_(False),
-    ).count()
-
-    purchases_count = db.query(TicketPurchase).filter(
-        TicketPurchase.user_id == user_id,
-        TicketPurchase.is_deleted.is_(False),
-    ).count()
-
-    scores_as_player1 = db.query(Score).filter(
-        Score.player1_id == user_id,
-        Score.is_deleted.is_(False),
-    ).count()
-
-    scores_as_player2 = db.query(Score).filter(
-        Score.player2_id == user_id,
-        Score.is_deleted.is_(False),
-    ).count()
-
-    total_scores = scores_as_player1 + scores_as_player2
-    can_delete = active_reservations == 0
-
-    recommendations = [
-        "Les scores seront conservés de manière anonymisée pour préserver l'intégrité des classements",
-        "Les réservations terminées seront préservées pour l'historique",
-        "Les données personnelles seront marquées comme supprimées conformément au RGPD",
-    ]
-    if not can_delete:
-        recommendations.append("⚠️ Annulez d'abord les réservations actives avant la suppression")
-
-    return {
-        "user": {
-            "id": user.id,
-            "pseudo": user.pseudo,
-            "email": user.email,
-            "tickets_balance": user.tickets_balance,
-            "created_at": user.created_at.isoformat(),
-        },
-        "can_delete": can_delete,
-        "blocking_factors": {
-            "active_reservations": active_reservations
-        } if not can_delete else {},
-        "deletion_impact": {
-            "friendships_to_delete": friendships_count,
-            "promo_uses_to_delete": promo_uses_count,
-            "purchases_to_delete": purchases_count,
-            "completed_reservations_preserved": completed_reservations,
-            "scores_anonymized": total_scores,
-        },
-        "recommendations": recommendations,
-    }
+    return get_user_deletion_impact_service(db, user_id)
 
 
 @router.put(
@@ -783,74 +203,14 @@ async def force_cancel_user_reservations(
     user_id: int,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Force l'annulation de toutes les réservations actives d'un utilisateur (admin uniquement)."""
-    from app.models.reservation import Reservation, ReservationStatus
-
-    user = db.query(User).filter(
-        User.id == user_id,
-        User.is_deleted.is_(False),
-    ).first()
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Utilisateur non trouvé",
-        )
-
-    active_reservations = db.query(Reservation).filter(
-        (Reservation.player_id == user_id) | (Reservation.player2_id == user_id),
-        Reservation.status.in_([ReservationStatus.WAITING, ReservationStatus.PLAYING]),
-        Reservation.is_deleted.is_(False),
-    ).all()
-
-    cancelled_count = 0
-    refunded_tickets = 0
-
-    for reservation in active_reservations:
-        reservation.status = ReservationStatus.CANCELLED
-
-        if reservation.player_id == user_id:
-            user.tickets_balance += reservation.tickets_used
-            refunded_tickets += reservation.tickets_used
-
-        cancelled_count += 1
-
-    db.commit()
-
-    return {
-        "message": f"Réservations de l'utilisateur '{user.pseudo}' annulées",
-        "user_id": user.id,
-        "cancelled_reservations": cancelled_count,
-        "refunded_tickets": refunded_tickets,
-        "new_tickets_balance": user.tickets_balance,
-    }
+    return force_cancel_user_reservations_service(db, user_id)
 
 
-# === STATISTIQUES ===
 @router.get("/stats")
 async def get_admin_stats(
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Récupère les statistiques globales de la plateforme."""
-    active_users = db.query(User).filter(User.is_deleted.is_(False)).count()
-    total_arcades = db.query(Arcade).filter(Arcade.is_deleted.is_(False)).count()
-    total_games = db.query(Game).filter(Game.is_deleted.is_(False)).count()
-    active_promo_codes = db.query(PromoCode).filter(
-        PromoCode.is_deleted.is_(False)
-    ).count()
-
-    total_tickets = db.query(func.sum(User.tickets_balance)).filter(
-        User.is_deleted.is_(False)
-    ).scalar() or 0
-
-    return {
-        "active_users": active_users,
-        "total_arcades": total_arcades,
-        "total_games": total_games,
-        "active_promo_codes": active_promo_codes,
-        "total_tickets_in_circulation": total_tickets,
-        "timestamp": datetime.now(timezone.utc).isoformat(),
-    }
+    return get_admin_stats_service(db)
 
 
 @router.delete("/arcades/{arcade_id}", responses=ARCADE_NOT_FOUND_RESPONSE)
@@ -858,64 +218,14 @@ async def soft_delete_arcade(
     arcade_id: int,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Supprime une borne d'arcade (soft delete)."""
-    from app.models.reservation import Reservation, ReservationStatus
-
-    arcade = db.query(Arcade).filter(
-        Arcade.id == arcade_id,
-        Arcade.is_deleted.is_(False),
-    ).first()
-
-    if not arcade:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=ARCADE_NOT_FOUND,
-        )
-
-    active_reservations = db.query(Reservation).filter(
-        Reservation.arcade_id == arcade_id,
-        Reservation.status.in_([ReservationStatus.WAITING, ReservationStatus.PLAYING]),
-        Reservation.is_deleted.is_(False),
-    ).count()
-
-    if active_reservations > 0:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Impossible de supprimer la borne : {active_reservations} réservation(s) active(s)",
-        )
-
-    now = datetime.now(timezone.utc)
-    arcade.is_deleted = True
-    arcade.deleted_at = now
-
-    arcade_games = db.query(ArcadeGame).filter(
-        ArcadeGame.arcade_id == arcade_id,
-        ArcadeGame.is_deleted.is_(False),
-    ).all()
-
-    for ag in arcade_games:
-        ag.is_deleted = True
-        ag.deleted_at = now
-
-    db.commit()
-
-    return {
-        "message": f"Borne '{arcade.nom}' supprimée avec succès",
-        "arcade_id": arcade.id,
-        "deleted_associations": len(arcade_games),
-    }
+    return soft_delete_arcade_service(db, arcade_id)
 
 
 @router.get("/arcades/deleted")
 async def list_deleted_arcades(
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Liste les bornes d'arcade supprimées (soft delete)."""
-    deleted_arcades = db.query(Arcade).filter(
-        Arcade.is_deleted.is_(True)
-    ).order_by(Arcade.deleted_at.desc()).all()
-
-    return deleted_arcades
+    return list_deleted_arcades_service(db)
 
 
 @router.put("/arcades/{arcade_id}/restore", responses=ARCADE_NOT_FOUND_RESPONSE)
@@ -923,76 +233,7 @@ async def restore_arcade(
     arcade_id: int,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Restaure une borne d'arcade supprimée."""
-    arcade = db.query(Arcade).filter(Arcade.id == arcade_id).first()
-
-    if not arcade:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Borne d'arcade non trouvée",
-        )
-
-    if not arcade.is_deleted:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Cette borne n'est pas supprimée",
-        )
-
-    existing_api_key = db.query(Arcade).filter(
-        Arcade.api_key == arcade.api_key,
-        Arcade.is_deleted.is_(False),
-        Arcade.id != arcade_id,
-    ).first()
-
-    if existing_api_key:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=(
-                "La clé API de cette borne est maintenant utilisée par une autre borne. "
-                "Veuillez générer une nouvelle clé API."
-            ),
-        )
-
-    arcade.is_deleted = False
-    arcade.deleted_at = None
-
-    arcade_games = db.query(ArcadeGame).filter(
-        ArcadeGame.arcade_id == arcade_id,
-        ArcadeGame.is_deleted.is_(True),
-    ).all()
-
-    restored_associations = 0
-    for ag in arcade_games:
-        game_exists = db.query(Game).filter(
-            Game.id == ag.game_id,
-            Game.is_deleted.is_(False),
-        ).first()
-
-        if game_exists:
-            slot_conflict = db.query(ArcadeGame).filter(
-                ArcadeGame.arcade_id == arcade_id,
-                ArcadeGame.slot_number == ag.slot_number,
-                ArcadeGame.is_deleted.is_(False),
-                ArcadeGame.id != ag.id,
-            ).first()
-
-            if not slot_conflict:
-                ag.is_deleted = False
-                ag.deleted_at = None
-                restored_associations += 1
-
-    db.commit()
-
-    return {
-        "message": f"Borne '{arcade.nom}' restaurée avec succès",
-        "arcade_id": arcade.id,
-        "restored_associations": restored_associations,
-        "note": (
-            f"{len(arcade_games) - restored_associations} association(s) non restaurée(s) en raison de conflits"
-            if restored_associations < len(arcade_games)
-            else None
-        ),
-    }
+    return restore_arcade_service(db, arcade_id)
 
 
 @router.put(
@@ -1003,26 +244,4 @@ async def regenerate_arcade_api_key(
     arcade_id: int,
     db: Annotated[Session, Depends(get_db)],
 ):
-    """Régénère la clé API d'une borne d'arcade."""
-    arcade = db.query(Arcade).filter(
-        Arcade.id == arcade_id
-    ).first()
-
-    if not arcade:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Borne d'arcade non trouvée",
-        )
-
-    old_api_key = arcade.api_key
-    new_api_key = generate_unique_arcade_api_key(db)
-
-    arcade.api_key = new_api_key
-    db.commit()
-
-    return {
-        "message": f"Clé API de la borne '{arcade.nom}' régénérée",
-        "arcade_id": arcade.id,
-        "old_api_key": old_api_key[:20] + "...",
-        "new_api_key": new_api_key,
-    }
+    return regenerate_arcade_api_key_service(db, arcade_id)
