+++ app/api/v1/games.py
@@ -1,10 +1,10 @@
-from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
-from typing import Annotated, List, Optional
+from typing import Annotated, List
+
 from app.core.database import get_db
-from app.models.game import Game
 from app.schemas.game import GameResponse
-from app.core.messages import GAME_NOT_FOUND
+from app.services.games_service import GameService
 
 router = APIRouter()
 
@@ -13,31 +13,14 @@ router = APIRouter()
 async def get_games(
     db: Annotated[Session, Depends(get_db)]
 ):
-    """Récupère la liste de tous les jeux disponibles."""
-
-    games = db.query(Game).filter(
-        Game.is_deleted == False
-    ).all()
-
-    return games
+    service = GameService(db)
+    return service.get_games()
 
 
 @router.get("/{game_id}", response_model=GameResponse)
 async def get_game(
-        game_id: int,
-        db: Annotated[Session, Depends(get_db)]
+    game_id: int,
+    db: Annotated[Session, Depends(get_db)]
 ):
-    """Récupère les détails d'un jeu spécifique."""
-
-    game = db.query(Game).filter(
-        Game.id == game_id,
-        Game.is_deleted == False
-    ).first()
-
-    if not game:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=GAME_NOT_FOUND
-        )
-
-    return game
+    service = GameService(db)
+    return service.get_game(game_id)
